Add a container-based development environment
There is now a Dockerfile that installs btfm's dependencies and
builds it, runs its tests, and builds its docs (the last three
enable the resulting container to cache all the built
dependencies).

There are also some helper scripts that use podman for common
tasks, such as building the image, running cargo, or running the
tests.

Signed-off-by: Randy Barlow <randy@electronsweatshop.com>