web: Adjust the URL routing to work with the CLI
Prior to axum v0.6 a route with a trailing slash also matched a request
without the slash. The CLI doesn't use a slash for resources so when the
server was updated commands on a single resource broke.

Remove the trailing slash for routes that deal with a single item rather
than a collection.