Move whisper-server into the btfm service
It was made as a separate service because that's how it all used to
work. There's not much reason to keep it that way, so move the
transcription into btfm to keep things easier to run. Also switch from
writing temporary files to converting the audio to Whisper's target
format (a numpy array of f32le 16khz mono audio)